Update Android implementation for Capacitor 4, rename module
