+++ android/src/main/java/com/joinflux/flux/segment/SegmentPlugin.java
@@ -1,23 +1,23 @@
-package com.joinflux.flux.capacitorsegment;
+package com.joinflux.flux.segment;
 
+import android.content.Context;
+import android.util.Log;
 import com.getcapacitor.JSObject;
 import com.getcapacitor.Plugin;
 import com.getcapacitor.PluginCall;
 import com.getcapacitor.PluginMethod;
 import com.getcapacitor.annotation.CapacitorPlugin;
-
-import android.content.Context;
-import android.util.Log;
 import com.segment.analytics.Analytics;
 import com.segment.analytics.Analytics.Builder;
 import com.segment.analytics.Options;
 import com.segment.analytics.Properties;
 import com.segment.analytics.Traits;
 
-@CapacitorPlugin(name = "CapacitorSegment")
+@CapacitorPlugin(name = "Segment")
 public class CapacitorSegmentPlugin extends Plugin {
+
     private boolean initialized = false;
-    private Segment implementation = new Segment();
+    private com.joinflux.flux.capacitorsegment.Segment implementation = new com.joinflux.flux.capacitorsegment.Segment();
 
     @PluginMethod
     public void initialize(PluginCall call) {
@@ -26,24 +26,27 @@ public class CapacitorSegmentPlugin extends Plugin {
             return;
         }
         String key = call.getString("key");
+        Boolean trackLifecycle = call.getBoolean("trackLifecycle", false);
         if (key == null) {
             call.reject("Write key is required to initialize plugin");
             return;
         }
 
         Context context = this.getContext();
-        Builder builder = new Analytics.Builder(context, key);
-        Boolean trackLifecycle = call.getBoolean("trackLifecycle", false);
-        if (trackLifecycle) {
-            builder.trackApplicationLifecycleEvents().experimentalUseNewLifecycleMethods(false);
-        }
-
-        Boolean recordScreenViews = call.getBoolean("recordScreenViews", false);
-        if (recordScreenViews) {
-            builder.recordScreenViews();
-        }
-        initialized = true;
-        implementation.analytics = builder.build();
+        this.getActivity()
+            .runOnUiThread(
+                new Runnable() {
+                    @Override
+                    public void run() {
+                        Builder builder = new Analytics.Builder(context, key);
+                        if (trackLifecycle) {
+                            builder.trackApplicationLifecycleEvents();
+                        }
+                        initialized = true;
+                        implementation.analytics = builder.build();
+                    }
+                }
+            );
         call.resolve();
     }
